commit 8, Some Validation done, some... new custom exceptions...

+++ src/main/java/edu/wctc/apw/apwmidtermapp/model/WineService.java
@@ -5,7 +5,9 @@
  */
 package edu.wctc.apw.apwmidtermapp.model;
 
+import edu.wctc.apw.apwmidtermapp.exception.DaoIsNullException;
 import edu.wctc.apw.apwmidtermapp.exception.DatabaseAccessException;
+import edu.wctc.apw.apwmidtermapp.exception.QueryParameterMissingException;
 import java.io.Serializable;
 import java.util.List;
 import javax.enterprise.context.SessionScoped;
@@ -21,23 +23,51 @@ public class WineService implements Serializable {
     @Inject
     private WineListDaoStrategy dao;
 
+    
+    
     public WineService() {
     }
     
-    public  Wine getWineById(String wineId) throws DatabaseAccessException {
+    
+    
+    public  Wine getWineById(String wineId) throws DatabaseAccessException, QueryParameterMissingException {
+        if(wineId.isEmpty() || wineId == null || Integer.parseInt(wineId) < 0){
+            throw new QueryParameterMissingException();
+        }
+        
         return dao.getWineById(Integer.parseInt(wineId));
     }
+    
+    
     
     public  List<Wine> getAllWines() throws DatabaseAccessException {
         return dao.getWineList();
     }
     
+    
+    
     public  void deleteWineById(String wineId) throws DatabaseAccessException {
+         if(wineId.isEmpty() || wineId == null || Integer.parseInt(wineId) < 0){
+            throw new QueryParameterMissingException();
+        }       
         dao.deleteWineById(Integer.parseInt(wineId));
     }
     
+    
+    
     public  void saveOrUpdateWine(String wineId, String wineName, double price, String imageUrl ) throws DatabaseAccessException {
-        
+         
+        // this first one might present a problem with nulls.  TEST IT.
+        if(Integer.parseInt(wineId) < 0){
+            throw new QueryParameterMissingException();
+        } if(wineName.isEmpty() || wineName == null ) {
+            throw new QueryParameterMissingException();
+        } if(price < 0 || price > 500){
+            throw new QueryParameterMissingException();
+        } if(imageUrl.isEmpty() || imageUrl == null ) {
+            throw new QueryParameterMissingException();
+        }
+         
         
         Integer id = null;
         if(wineId == null || wineId.isEmpty()){
@@ -54,6 +84,9 @@ public class WineService implements Serializable {
     }
 
     public void setDao(WineListDaoStrategy dao) {
+        if(dao == null){
+            throw new DaoIsNullException();
+        }
         this.dao = dao;
     }
